Generics changes for o.a.c.connector Fix various Eclipse warnings (unused code, casts, etc)
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@729651 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/catalina/connector/CometEventImpl.java
@@ -26,7 +26,6 @@ import javax.servlet.http.HttpServletResponse;
 
 import org.apache.catalina.CometEvent;
 import org.apache.catalina.util.StringManager;
-import org.apache.coyote.ActionCode;
 
 public class CometEventImpl implements CometEvent {
 
@@ -119,7 +118,7 @@ public class CometEventImpl implements CometEvent {
             UnsupportedOperationException {
         if (request.getAttribute("org.apache.tomcat.comet.timeout.support") == Boolean.TRUE) {
             request.setAttribute("org.apache.tomcat.comet.timeout", new Integer(timeout));
-            if (request.isComet()) request.setCometTimeout((long)timeout);
+            if (request.isComet()) request.setCometTimeout(timeout);
         } else {
             throw new UnsupportedOperationException();
         }
